[starnix] Refactor task data needed to compute what process can be waited
Before this CL, the ThreadGroup contained a Map of Weak<Task> and was
considering that these weak pointers could always be upgraded because
they were cleared when the Task was dropped. This is incorrect, because
a Weak pointer will become not upgradable before the drop implementation
is run. To fix this race, this CL does the following:
1. Stop considering the weak pointer is always upgradable
2. For computing whether a process is a target for a given wait, share
   some data with the tasks, so that this data is available even when
   the task destruction is in flight.

Change-Id: I4e52f3cbe0bc79fddf21c3bfac4cebb96cf10a70
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/878491
Fuchsia-Auto-Submit: Benjamin Lerman <qsr@google.com>
Reviewed-by: Kevin Lindkvist <lindkvist@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>