[starnix] Run syscall tests on host
Involves:

a) Skipping a number of tests that require specific
capabilities if those capabilities are not present on the
system.

b) Ensuring that tests that only need to wait for specific
process children don't wait for all process children (and updating
ForkHelper to allow clients to do that).

c) Making some tests starnix-only when I didn't immediately
know what was causing them to fail on Linux in CQ, or
didn't otherwise see an obvious workaround.

Bug: 129749

Change-Id: Ibc78a88761b0af45832f6ca8f784383a4b5fe8b1
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/875719
Reviewed-by: Kevin Lindkvist <lindkvist@google.com>
Commit-Queue: Jeremy Manson <jeremymanson@google.com>