[packet] Simplify execution of Serializers
Previously, `Serializer::serialize` consumed a `NestedPacketBuilder`
which represented the layers outside of `self`. The innermost call to
`serialize` was responsible not only for allocating a buffer which
satisfies the various constraints (especially prefix and suffix length),
but was also responsible for performing serialization itself by
delegating to the `NestedPacketBuilder` provided to it as an argument.
As a result of this design, serialization was more complicated in terms
of code and in terms of call paths than necessary.

In this commit, we simplify `serialize` in the following ways:
- We remove `NestedPacketBuilder`
- `serialize` takes a `PacketConstraints` which represents the outer
  layers, but does not take any `PacketBuilder` or `NestedPacketBuilder`
- The innermost call to `serialize` is still responsible for allocating
  a buffer which satisfies the relevant constraints, but is no longer
  responsible for serializing outer layers
- Instead, on the return path, `serialize` is responsible only for
  serializing the `PacketBuilder` owned by `self` - all encapsulated
  layers have already been serialized by recursive calls to `serialize`,
  and all outer layers will be serialized in the future by callers

In practice, this change results in a significant reduction in LoC in
the `serialize` module.

Note that the API consumed by most users - namely methods on
`Serializer` with an `_outer` suffix - is unaffected by this change.

Change-Id: I725551a3dbba853f12358113e056d9695b93ec58
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/861377
Reviewed-by: Alex Konradi <akonradi@google.com>
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>
Reviewed-by: Joshua Liebow-Feeser <joshlf@google.com>
Commit-Queue: Joshua Liebow-Feeser <joshlf@google.com>