[network/tests] Disable NS2 tests in generated src
Include the `#[cfg(not(target_arch = "riscv64"))]` attribute in the
`#[netstack_test]` macro output, not in the source code for the macro.
Since the macro is compiled on the host machine, it isn't aware of what
architecture its generated output is going to be compiled for. The
attribute needs to be present in the macro's generated output so that
when that output is compiled for the target system, it can disable some
test functions for RISC-V.

To demonstrate the fix and prevent backsliding, add some network tests
to the test group run for the minimal product, which will be run on
RISC-V.

Bug: 122609
Fixed: 129748
Change-Id: I2870c6e0627ce2ede58f072eaeeb1dc9b1f1c476
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/878477
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Nick Brown <nickbrow@google.com>
Fuchsia-Auto-Submit: Alex Konradi <akonradi@google.com>
Reviewed-by: Aaron Wood <aaronwood@google.com>