[boot-shim] Introduce RiscvDevicetreePlicItem
This item is in charge of parsing the plic node from the devicetree.
For now it only parses the base address and the number of IRQs, which is
enough for what the current zbi plic driver needs.

As part of this change, PropertyDecoder has been extended with
|TranslateAddress| method which recursively translates addresses
until is not able to. This is necessary for all addresses, and its been
observed in devices dts. Since in order for the driver to properly
understand the mmio address it needs the view of the root node of the
memory address. So it needs to be translated all the way to the root if
possible.

Bug: 129970
Test: fx test -v -o boot-shim-unittests
Change-Id: I8be362de417cf08e2f8beda9d4f91c42e55179f5
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/879511
Reviewed-by: Joshua Seaton <joshuaseaton@google.com>
Commit-Queue: Gianfranco Valentino <gevalentino@google.com>