[rbe][rust] Expand linker scripts
Remote linking (e.g. for Rust) may encounter link args
that are linker scripts, as opposed to archives/binaries.
Linker scripts need to be evaluated to determine
all of the files that will be needed for the eventual link.

With this, we can un-hardcode some of the sysroot files
listed in fuchsia.py, and make it more robust to sysroot updates.

Bug: 129842
Change-Id: I76d5fae9c0f550248adfbdc2eacada91cc7dd42a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/879972
Reviewed-by: Jay Zhuang <jayzhuang@google.com>
Fuchsia-Auto-Submit: David Fang <fangism@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>