[swd] Prepare for moving system-updater
I will move `/core/system-updater` to
`/core/system-update/system-updater` which will break this tool
and it is not possible with our current roller strategy to roll
`ffx` and the updated system image together. Luckily `/core`
already exposes the capability we need, so we can connect to it
there until the transition is complete.

Bug: 123798
Change-Id: Ie07c3e68edf5054c4cc7ba8101f3a5ae4414eee5
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/880031
Commit-Queue: Justin Mattson <jmatt@google.com>
Reviewed-by: Aaron Wood <aaronwood@google.com>
Reviewed-by: Megan Batty <mgnb@google.com>