[superproject] Roll scudo [scudo] Extract steps releaseToOSMaybe into functions in
This refactor helps us identify which steps need FLLock so that we can
reduce the holding time of FLLock in SizeClassAllocator64.

Also move the data members to the end of class to align the style in
SizeClassAllocator32.

Reviewed By: cferris

Differential Revision: https://reviews.llvm.org/D152596

GitOrigin-RevId: e2483abaadddb9cd1361ad68d35485f89a9dba48
Original-Revision: e00d761b5a7c08a2dcee5cd8991ce7171fe39a34
Roller-URL: https://ci.chromium.org/b/8776363213821731777
CQ-Do-Not-Cancel-Tryjobs: true
Change-Id: I653f3bb33dcecf0124bac66b31b993b3f71f5c05
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/881242