[sparse] Separate sparse to a library
Allows code handling the sparse format to be reused (in particular, in
the paver).

The interface is slightly different from the interface in
firmware/lib/storage, in that there are no GPT details involved in the
interface, and a generic handle type is used for I/O rather than raw
buffers.  This is more flexible and accommodates the paver's use-case,
where the paver will write directly to a block device representing a
single partition in the GPT.

The interface in firmware/lib/storage is now a thin shim over the
sparse library.

Bug: 129533
Change-Id: Ifdb0c4abdd392863af43776245d8787a666ba8ae
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/878652
Reviewed-by: Megan Batty <mgnb@google.com>
Commit-Queue: James Sullivan <jfsulliv@google.com>
Reviewed-by: Dov Shlachter <dovs@google.com>
Reviewed-by: David Pursell <dpursell@google.com>