[microbenchmarks] Add Vmo/Memcpy/WithPrecommit benchmarks
Adds some microbenchmarks that measure the time it takes to copy various
amounts of data into a VMO, either with or without committing the pages
in the destination prior to copying. This is an indirect way of
measuring the overhead induced by page faulting during a copy.

Change-Id: I473da2609a75ff5bc65f1822afdb4fd78f1fdc5d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/858406
Reviewed-by: Mark Seaborn <mseaborn@google.com>
Commit-Queue: Anirudh Mathukumilli <rudymathu@google.com>
Fuchsia-Auto-Submit: Anirudh Mathukumilli <rudymathu@google.com>