[vfs] Apply the typestate pattern [1] on node names
fuchsia.io/Name has a lot of constraints, but not all of them are
respected in the VFS. There are inconsistent checks in the `simple`
directory and path parsing as to what constitutes a valid filesystem
name.

This CL introduces a `Name` type, which wraps a String, and can only be
constructed if the string satisfies the conditions for a valid name, as
spelled out in the fuchsia.io spec. This way as long as we see a `Name`
object we know the name must be valid, and does not have to check again.

This CL updates the DirectlyMutable trait to take `Name`. I couldn't
update `TraversalPosition::Name` because of a non-fuchsia.git repository
depending on it.

[1]: http://cliffle.com/blog/rust-typestate/

Bug: 130228

Change-Id: I98444258e041cc42858e5ce0844adc147318d928
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/881213
Reviewed-by: James Sullivan <jfsulliv@google.com>
Reviewed-by: Didi She <didis@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Yifei Teng <yifeit@google.com>