[memalloc] Introduce Pool::NormalizeRanges()
This utility provides a callback with a normalized view of all tracked
ranges, reducing and merging types as desired. In mapping in the
entirety of RAM in physboot, it is useful to generate a normalized view
of RAM, removing any distinction between free and allocated RAM: we
update one site to use this method where we hand-roll this logic today.

Change-Id: Ibbcf7b680406da7f36dd94dbd2c0c10b7bde79d0
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/880981
Commit-Queue: Joshua Seaton <joshuaseaton@google.com>
Fuchsia-Auto-Submit: Joshua Seaton <joshuaseaton@google.com>
Reviewed-by: Roland McGrath <mcgrathr@google.com>