[fdf] Fully support specs in ffx driver list-composites
Update ffx driver list-composites so that it contain all
composite node specs, including the ones not yet matched
to a driver, and is supported in DFv1 and DFv2.

The list of CompositeInfo from specs is provided by
CompositeNodeSpecManager.

This CL also refactors ParentSetCollector so that it
contains the composite driver information and is
responsible for assembling the composite node.

Bug: 119947
Tests: driver-runner-tests
Change-Id: I65f4bc37732fbad6487fae47ccb8537e482f822e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/877417
Reviewed-by: Novin Changizi <novinc@google.com>
Commit-Queue: Sarah Chan <spqchan@google.com>