[tracing] Add a core shard target
As part of migrating tracing to be fully optional, it is time to remove the
core shard from non eng builds. However, out of tree defined products
still may depend on the core shard so they need a way to depend on it.

We can't add it first because we get duplicate includes.
We can't remove it first because it will temporarily not be defined.

So we set up a migration where first we land this patch.
Then we point the out of tree products to use this empty target.
Then in the following patch, we simultaneous remove the core shard and
add it to this target.

Bug: 112433
Change-Id: I4ac556a21a4f971422fba064ef63162e3afaf717
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/881768
Reviewed-by: Brian Hamrick <bhamrick@google.com>
Fuchsia-Auto-Submit: Gwen Mittertreiner <gmtr@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>