[driver_runner] Remove recursion in ScheduleStopComponent
We observe the following sequence of calls can occur resulting in
recursion:

-> CheckForRemoval
  -> ScheduleStopComponent
    -> FinishRemoval
      -> RemoveChild
        -> CheckForRemoval

ScheduleStopComponent only eagerly calls FinishRemoval in the case that
the driver component is not there. By posting an async task to call it
later we break the recursion cycle and ensure the same flow of is
observed in both the optimistic case as well as the slow case.

Additionally, a number of clang tidy lints were addressed.

Test: fx test driver-runner-test.cm
Change-Id: I3823463216b0e928736efbb3f8c51255d57e1298
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/882586
Fuchsia-Auto-Submit: Suraj Malhotra <surajmalhotra@google.com>
Reviewed-by: Sarah Chan <spqchan@google.com>
Reviewed-by: Novin Changizi <novinc@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>