[superproject] Roll llvm-project/libc [libc] Fix using the `libcgpu.a` for NVPTX in non-LTO builds
CUDA requires a PTX feature to be compiled generally, because the
`libcgpu.a` archive contains LLVM-IR we need to have one present to
compile it. Currently, the wrapper fatbinary format we use to
incorporate these into single-source offloading languages has a special
option to provide this. Since this was not present in the builds, if the
user did not specify it via `-foffload-lto` it would not compile from
CUDA or OpenMP due to the missing PTX features. Fix this by passing it
to the packager invocation.

Reviewed By: jdoerfert

Differential Revision: https://reviews.llvm.org/D154864

GitOrigin-RevId: da88537c54fd3a482232117f64f3ec9e16d6e1e0
Original-Revision: 4423547b7b6462cba78baf09c4b3080d121f4f7e
Roller-URL: https://ci.chromium.org/b/8775924718669613409
CQ-Do-Not-Cancel-Tryjobs: true
Change-Id: I6ae673ea066df32282e5fbce716dddbfed30ed43
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/882691