[assembly] Images manifest uses namespace not name
When we collect the GN metadata for the images.json file, we previously
used the image_name specified for assembled_system(). This works great
when image_name is defaulted to the same thing as target_name. It does
not work well when image_name is specified to be "fuchsia", because we
end up with collisions in our images.json.

Usually when we override image_name and set it equal to "fuchsia" it is
because we are trying to reuse an existing images config. In those
situations we use namespace to keep the same output directory name as
before. Here is an example:

  assembled_system() {
    namespace = "netboot"
    image_name = "fuchsia"
    assembly_image_config_label = "//boards/images:default"
    ...
  }

The zbi and vbmeta will be generated in these locations:
  $target_out_dir/netboot/fuchsia.zbi
  $target_out_dir/netboot/fuchsia.vbmeta

We will use namespace instead of image_name from now on, so that we
avoid conflicts in our images.json.

Bug: b/290251101
Change-Id: I5d46302be24e22d95bce297fdc55d824baf151f3
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/882912
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Aidan Wolter <awolter@google.com>
Reviewed-by: David Iserovich <diserovich@google.com>