[wlan] Fix tests that are sensitive to the executor
I have an upcoming change that impacts tests that are sensitive to the
order in which the executor polls things. This fixes the tests in a way
that allows them to pass before and after the change.

Bug: 130421
Change-Id: I796614e6ebef76a074a4f2a7d8fc3464c2da913f
Multiply: wlancfg-tests
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/882857
Reviewed-by: Marc Khouri <mnck@google.com>
Fuchsia-Auto-Submit: Chris Suter <csuter@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>