[starnix] Access current task's memory through CurrentTask
This updates most callsites that are accessing memory associated with
the currently executing task to use a MemoryAccessor implementation
directly on CurrentTask instead of accessing the task's MemoryManager.
This is preparation for accessing this memory through mappings directly
in the future. Accesses to non-current tasks will still go through the
MemoryManager.

This also updates the parameter name used for CurrentTask in many
functions to distinguish it more clearly from a Task that may be
associated with a different thread.

Change-Id: I0ab248ba037a0850768b9763cb58734742a75737
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/881114
Commit-Queue: James Robinson <jamesr@google.com>
Reviewed-by: Kevin Lindkvist <lindkvist@google.com>