[Starnix] Add support for test arguments in LTP test runner
Previously all LTP tests were executed as a single binary. This change
allows to specify test binary that's different from the test name and
pass arguments when necessary.
Now the tests list file is interpreted as LTP run files, see
https://github.com/linux-test-project/ltp/blob/master/doc/ltp-run-files.txt .
Each line contains the name of the test as the first word, followed
by the script to run to execute the test. Lines with a single word are
handled the same way as before - i.e. test name is used as the name of
the binary to run.

Also, removed support for test binary enumeration when there is no
test list file.

Bug: b/290795980
Change-Id: I6dfc613eaa55c9acd17c936af35a88f6b58538fd
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/883119
Commit-Queue: Sergey Ulanov <sergeyu@google.com>
Reviewed-by: Kevin Lindkvist <lindkvist@google.com>