[debugger] Get return values of built-in types from function calls
This change will allow for functions with return values to display the
value at the end of the function call, or interact with the return value
in a greater expression.

Note: Some user defined POD types may actually work with this patch
(specifically if they can fit into registers, per the particular ABI of
the target), but generally user defined types should not be expected to
work.

Bug: 5457
Change-Id: I32b13f5c404965171add57d793b84c3fb2d2f045
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/879419
Commit-Queue: Jacob Rutherford <jruthe@google.com>
Reviewed-by: Dangyi Liu <dangyi@google.com>