[riscv64] Fix ABI issue in riscv platform.cc
uint32_t Foo was defined instead of ktl::optional<uint32_t>.
So the value was returned in one register and garbage on the other, the
side effect is that the value is correct but the flag being zero, caused
it to have a valid value and but !value().

Now we address this by adding the proper include and fixing the return
type.

Change-Id: I6d63aaeaa273c2bc575685fa6110eebd9fa41372
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/883584
Commit-Queue: Gianfranco Valentino <gevalentino@google.com>
Reviewed-by: Roland McGrath <mcgrathr@google.com>